duplicate function in query deleted.
